requirements.txt
bibtexparser==1.4.1
biopython==1.81
crispy-bulma==0.10.0
django-crispy-forms~=2.0
django-environ==0.11.2
django-filter==23.3
Django==4.2.6
djangorestframework==3.14.0
djangorestframework_simplejwt==5.3.1
django-cors-headers==4.6.0
fasttext~=0.9.2
fuzzywuzzy==0.18.0
grobid_tei_xml==0.1.3
markdown==3.5
numpy==1.26.1
pandas==2.1.1
psycopg2-binary~=2.9.3
rispy==0.8.1
rdflib==7.0.0
requests==2.31.0
scholarly~=1.7.2
wikipedia~=1.4.0
xmltodict~=0.13.0